+++ NEWS.md
@@ -1,7 +1,7 @@
 # SoilTaxonomy 0.2.2 (2022-10-01)
  * The order of dataset `ST` is now based on the full subgroup level code (https://github.com/ncss-tech/SoilTaxonomy/issues/35). Partial codes corresponding to each level of the hierarchy are also included in columns: `order_code`, `suborder_code`, `greatgroup_code` and `subgroup_code`.
  * New function `SoilTaxonomyLevels()`: used for creating ordinal and nominal factors corresponding to taxonomic levels `"order"`, `"suborder"`, `"greatgroup"`, and `"subgroup"` (https://github.com/ncss-tech/SoilTaxonomy/issues/39).
- * New function `SoilMoistureRegimeLevels()`: used for creating ordinal and nominal factors corresponding to Soil Moisture Regimes (https://github.com/ncss-tech/SoilTaxonomy/issues/34).
+ * New function `SoilMoistureRegimeLevels()` and `SoilTemperatureRegimeLevels()`: used for creating ordinal and nominal factors corresponding to Soil Moisture and Temperature Regimes (https://github.com/ncss-tech/SoilTaxonomy/issues/34).
  * `FormativeElements()` can now take multiple `level` values, returning the formative elements found at each level combined in a single data.frame.
  * New function `get_ST_formative_elements()` a helper function for getting a data.frame containing descriptors of formative elements used at the specified `level`(s)
  * New function `extractSMR()`: utilizes order to subgroup formative elements to identify Soil Moisture Regimes from taxon names